+++ scripts/content/lesson/lesson.js
@@ -113,6 +113,13 @@
 			});
 		});
 
+		// listen for the no challenge event
+		document.addEventListener("nochallenge", async event => {
+			if (typeof extensionActive == "function" && !(await extensionActive())) return; 
+
+			clearAll();	
+		});
+
 		document.addEventListener("click", async event => {
 			const target = event.target;
 			
@@ -148,6 +155,11 @@
 					
 					navigator.clipboard.writeText(JSON.stringify(data, null, 2));
 				}
+			}
+
+			// close speech bubble on click outside
+			if (!target.closest(".d-cgpt-speech-bubble") && document.querySelector(".d-cgpt-speech-bubble")?.style.display !== "none") {
+				toggleExtraInput(false);
 			}
 		});
 
@@ -514,10 +526,7 @@
 		button.classList.remove("d-cgpt-explain-button-enabled");
 		button.classList.add("d-cgpt-explain-button-disabled", "_1NM1Q");
 
-		const explainBubble = document.querySelector(".d-cgpt-speech-bubble");
-		if (explainBubble) {
-			explainBubble.style.display = "none";
-		}
+		toggleExtraInput(false);	
 	}
 
 	const enableButton = button => {
@@ -548,6 +557,23 @@
 				disableButton(button);
 			});
 		}
+
+		// right click
+		button.addEventListener("contextmenu", event => {
+			console.log("right click");
+			if (window.innerWidth <= 700) {
+				event.preventDefault();
+
+				const speechBubble = document.querySelector(".d-cgpt-speech-bubble");
+				if (speechBubble) {
+					toggleExtraInput(true);
+
+					speechBubble.classList.add("d-cgpt-speech-bubble-mobile");
+					speechBubble.focus();
+				}
+			}
+			
+		});
 		
 		button.dataset.hasclick = true;
 	}
@@ -592,22 +618,27 @@
 	const toggleExtraInput = show => {
 		const extraInput = document.querySelector(".d-cgpt-speech-bubble");
 		if (extraInput) {
+			const showBubble = () => {
+				extraInput.style.removeProperty("display");
+			}
+
+			const hideBubble = () => {
+				extraInput.style.display = "none";
+				extraInput.classList.remove("d-cgpt-speech-bubble-mobile");
+			}
+
 			if (show == null) {
-				if (extraInput.style.display == "none") {
-					extraInput.style.removeProperty("display");
-				}
-				else {
-					extraInput.style.display = "none";
-				}
+				if (extraInput.style.display == "none")
+					showBubble();
+				else
+					hideBubble();
 				return;
 			}
 
-			if (show) {
-				extraInput.style.removeProperty("display");
-			}
-			else {
-				extraInput.style.display = "none";
-			}
+			if (show)
+				showBubble();
+			else
+				hideBubble();
 		}
 	}